Jetty 9.2 upgrade follow-up fix
After running this new version of test harness with Jenkins core, I
discovered one issue around the authentication managed by servlet
container.

When Jenkins is configured to delegate authentication to servlet
container, Jetty 8.x used to allow any user, regardless of the roles
they have in Jetty, to qualify. This was because web.xml has the
following and Jetty interpreted '*' to "any role regardless of whether
the role is defined in this web.xml or not"

    <security-constraint>
        <web-resource-collection>
            <web-resource-name>Hudson</web-resource-name>
            <url-pattern>/loginEntry</url-pattern>
            <!--http-method>GET</http-method-->
        </web-resource-collection>
        <auth-constraint>
            <role-name>*</role-name>
        </auth-constraint>
    </security-constraint>

Apparently this was one of the ambiguity in the servlet spec, and as of
servlet 3.1, the spec has locked this down and resolved against what
Jetty used to do. Now role-name="*" means "any role defined in web.xml",
and defined role-name="**" to mean "any role".

The net result is that, with the user realm setup here, none of the
users can log on to Jenkins. So I'm adding back "user" role so that
these users pass role-name="*" check.

Jenkins 2 will use the new role-name="**" but that is not available in
servlet 2.5 that Jenkins 1.x uses.