Make dependency declaration less error prone
If a plugin explicitly declares a dependency on core, war, or some combination of it, but without declaring a dependency to war-for-test, unit tests end up running with one version of the core jar file and another version of the war file. This results in a strange error that's hard to diagnose.

It's very unlikely that plugins using test harness just so happens to be based on 1.580.1, so it's better not to let this module brings in war-for-test, and instead let the caller declare it explicitly, like the plugin parent POM is doing.