use a subdirectory of the temp directory by default.
1) we are Jenkins :-)
2) put all the files in a directory inside temp rather than directly in
temp.

The problem with using the tempDirectory directly is that you can not add
exclusions with wildcards in many anti virus products.  by using a
subdirectory you can add an exclusion to your PATH_TO_TEMP\jenkinsTests
which will then allow you to exclude jenkinsTests but still allowing
scanning of your temporary directory (which is desired).