constrain an agents memory
Notivced when looking at the memory usage of pipeline-model-defintion
that the agents created in the tests did not have a constrained memory
set.  This means that the tests would be platform dependent. or flaky
depending on what GC happens at various times.

the number 512 is relativly arbitrary, but k8s agents use 256Mi as a
request and so 512 is double that, and the agent process is supposed to
be lightweight.

Whilst this could reveal a few issues in plugins when they upgrade, it
seems to me that that could be a bug in the plugin, and so I have not
implemented a way to create an online slave with a specific amount of
memory.  if that is indeed required it can be added later as jth updates
are not forced on plugins.