WarExploder: Handle path structures with a hash level
Currently WarExploder does not find the jenkins-core WAR file based on JAR file location if their directories contain an additional level for hash (checksum) of the file. This commit extends the search for such situations.

The problem currently bites me as I'm trying to cover a shared library with a test suite based on jenkins-pipeline-shared-libraries-gradle-plugin examples, in IntelliJ IDEA.