Test that read-only files are disposed
When the git plugin and the gitlab plugin upgraded to parent pom 4.64,
their tests failed on Windows. The Java delete() method on Windows
throws an access denied exception when asked to delete a read-only
file.

This test creates a read-only file and relies on the JenkinsRule
cleanup process to remove the file.