Accept NPE in MockFolder.createProject() if wrong job type is passed
The null pointer exception will happen when the test is run and will be
an immediate sign to the developer that there is a mistake in the caller.

Detected while prototyping an upgrade of the minimum Jenkins version
from 2.361.4 to 2.401.3.  Message is not reported with 2.361.4 as Jenkins
minimum version but is reported with Jenkins 2.387.3 or later.