Fix null incompatability
User.Get has a return of CheckForNull yet the returned value is passed
directly as a parameter of a method marked NonNull.

Whilst the return in this case can never be null as the create parameter
is true, compilers and tools are not clever enough to understand the
human written contract.  so we assert the user in non null before then
passing it down the line, to keep compilers happy