Update baseline to 2.361 to prevent detached plugins
the current baseline caused some detached plugins to be installed when
starting a RealJenkinsRule.

This has an undersirable effect that when
`includeTestClasspathPlugins(false)` is called we still end up with some
plugins that can negatively impact the test.

To avoid this we are simply moving the baseline forward to 2.361 (which
is what the j-t-h specifies as its Jenkins version!)