Add option to dynamically update the baseline of the RJRInit plugin
if the version of Jenkins used by RealJenkinsRuleInit has detached
plugins in the version of Jenknis used for the test then this can
uncessescarily pull in detached plugins.

In most cases this is harmless, but in some cases these plugins may be
prevented from running under the test (e.g. testing FIPS and one of the
detached plugins like eddsa is not FIPS complaint and blocks startup).

The plugin itself has limited exposure to the Jenkins api (via
hudson.plugin.Plugin and RealJenkinsRule$Init2).

Code running from the plugin tests inside RealJenkinsRule use the
uberClassloader and if the plugin was targetting a lower core (where the
dependencies have been deteached, they should still be included as this
change only affects the version of the RealJenkinsRuleInit plugin.